<INT> should return a number type compatible with matchers

+++ src/tailspin/interpreter/Composer.java
@@ -20,7 +20,7 @@ public class Composer implements Transform {
   private static final HashMap<String, Function<? super String, Object>> namedValueCreators = new HashMap<>();
   static {
     namedPatterns.put("INT", Pattern.compile("[+-]?(0|[1-9][0-9]*)"));
-    namedValueCreators.put("INT", Integer::valueOf);
+    namedValueCreators.put("INT", Long::valueOf);
     namedPatterns.put("WS", Pattern.compile("\\s+"));
     namedValueCreators.put("WS", Function.identity());
   }
